refactor: separate Main from JoinUser
